pyproject.toml
[project]
name = "rag-chatbot-api"
version = "0.1.0"
description = "RAG Chatbot API for querying book content with strict grounding in retrieved text"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.29.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.3.8",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.8.0",
    "requests>=2.31.0",
    "httpx>=0.28.1",
    "tiktoken>=0.5.0",
    "langchain>=0.0.335",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.0.5",
    "psycopg[binary]>=3.1.12",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0"
]
